Fixed burst limits
1. Firstly, the main link should always have enough tokens to cover full
burst speed, and cburst speed. If we don't we run into negatives and
traffic drops.

2. Add cburst. Set the class burst (and borrowing) to 10% of the difference
of Limit - CIR to allow a sort of fairness between classes based on how much
leeway is provided to utilize free bandwidth.

burst was set based on Limit, this is wrong. If mulitple classes Limits
add up to more than the CIR of the main link, the main link tokens will
be depleted. burst is now based on the CIR. cburst is now based on the
Limit.

3. Fixed the calculation of RED min and max to make the WMA algorith
work and not moan.